Relocate file and remove unused imports
