+++ src/components/Layout/Layout.tsx
@@ -2,7 +2,7 @@ import React, { FC } from "react"
 import Footer from "@/components/Footer/Footer"
 
 // Global state management
-import ContextTree from "../ContextTree"
+import ContextTree from "@/context/ContextTree"
 
 const Layout: FC = ({ children }) => {
   return (
